requirements.txt
asgiref==3.6.0
certifi==2022.12.7
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
gunicorn==20.1.0
mongomock==4.1.2
packaging==23.1
pymongo==4.3.3
pytz==2023.3
sentinels==1.0.0
sqlparse==0.4.4
typing_extensions==4.5.0
